Workshop > Craft Room
I'm personally on the fence about this, but consistency across game mechanics holds priority here. There are other instances of Workshop that fit naturally, like Theodore's Workshop. I also believe it is fine to introduce this into NPC text referring to the Craft Room. I.E. Sonia, the Craft Officer in the Craft Room is too redundant. Sonia, the Craft Officer in the workshop, for example, I'm good with, and would actively encourage if such a thing needs to be written.

If we want to elevate クラフトルーム Craft Room from its dogmatic bind dictated by its romaji, I would change it to Crafting Hall to sound more natural in English. 'Craft' is kept for quick visual indexing.