Go to the designated location... > Root out the rabid beasts
Go to the designated location and find the rabid demon > Root out the rabid beasts

Compressed the meaning of "go there" + "get rid of someone/something" (implied) with "Root out"

Changed demon > beasts as "mamono" [魔物] is a vague term meaning any sort of evil or dangerous creature, informed by the fact that these quests have you fight Wolves and Saurians in the pool of possible enemies. Did not change to "enemies" as that would be "teki" [敵] which is not used here.

Possible to add a (s) on beast(s) if ever there is just one enemy encountered.