[倒す] Defeat > Slay
We have two versions of "Defeat the encountered demons" that are translated the same.
[遭遇した魔物を討伐する]
[遭遇した魔物を倒す] << the subject of this commit
In fact there is a slight difference, the one using [討伐する] "tōbatsusuru" which means to subjugate or subdue in a military context, e.g. to vanquish a foe, and this one I am changing with [倒す] "taosu" is more of a simple kill order.

This doesn't have to be the final version, I'm changing it now so that it's easily differentiated for an English reader.
Personally, I would prefer if the former phrase [遭遇した魔物を討伐する] is translated to "Vanquish the encountered enemy" and then this latter one [遭遇した魔物を倒す] becomes Defeat the encountered enemy. This would be purposely overlooking  [魔物] "mamono" and changing it to "the target" as the player sees in game. This solution seems the most intuitive to me as these objective strings are part two of quests given by NPCs:
1. go find them (what+where?)
2. defeat them (here!)
World Quests will start on the defeat-upon-encounter objective.

In addition, the use of [魔物] "mamono" in both these phrases also refers to non-demons, so that will have to be changed later.